Feat: 패키지 구조 변경 및 라인 리팩토링
전체 패키지 구조를 domain 별로 변경하고 인텔리제이 기능을 사용해 라인을 리팩토링함
