+++ backend/WWMeet_backend/build.gradle
@@ -29,11 +29,6 @@ dependencies {
 	annotationProcessor 'org.projectlombok:lombok'
 	testImplementation 'org.springframework.boot:spring-boot-starter-test'
 	testImplementation group: 'com.h2database', name: 'h2', version: '2.1.214'
-
-	// map struct 의존성
-	implementation 'org.mapstruct:mapstruct:1.4.2.Final'
-	annotationProcessor "org.mapstruct:mapstruct-processor:1.4.2.Final"
-	annotationProcessor 'org.projectlombok:lombok-mapstruct-binding:0.2.0'
 }
 
 tasks.named('test') {
