+++ backend/WWMeet_backend/src/main/java/com/example/wwmeet_backend/appointment/dto/AppointmentRequestDto.java
@@ -1,0 +1,20 @@
+package com.example.wwmeet_backend.appointment.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.time.LocalDateTime;
+
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class AppointmentRequestDto {
+    private Long id;
+    private String appointmentName;
+    private String appointmentPlace;
+    private String identificationCode;
+    private int peopleNum;
+    private LocalDateTime deadline;
+}
