+++ backend/WWMeet_backend/src/main/java/com/example/wwmeet_backend/appointment/dto/SaveAppointmentResponseDto.java
@@ -1,0 +1,13 @@
+package com.example.wwmeet_backend.appointment.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class SaveAppointmentResponseDto {
+    private AppointmentResponseDto appointmentRequestDto;
+    private String participantName;
+}
