+++ backend/WWMeet_backend/src/main/java/com/example/wwmeet_backend/appointment/service/AppointmentService.java
@@ -1,0 +1,45 @@
+package com.example.wwmeet_backend.appointment.service;
+
+
+import com.example.wwmeet_backend.appointment.domain.Appointment;
+import com.example.wwmeet_backend.appointment.repository.AppointmentRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.NoSuchElementException;
+import java.util.Optional;
+
+@Service
+@RequiredArgsConstructor
+@Transactional(readOnly = true)
+public class AppointmentService {
+    private final AppointmentRepository appointmentRepository;
+    public List<Appointment> findAllAppointment(List<String> appointmentCodeList) {
+        List<Appointment> findAppointmentList = new ArrayList<>();
+        for (String appointmentCode : appointmentCodeList) {
+            Optional<Appointment> findAppointmnetOptional = appointmentRepository.findByIdentificationCode(appointmentCode);
+            Appointment findAppointment = findAppointmnetOptional
+                    .orElseThrow(() -> new NoSuchElementException());
+            findAppointmentList.add(findAppointment);
+        }
+        return findAppointmentList;
+    }
+
+    public Appointment findAppointmentById(Long id){
+        Optional<Appointment> findAppointmentOptional = appointmentRepository.findById(id);
+        return findAppointmentOptional
+                .orElseThrow(() -> new NoSuchElementException());
+    }
+
+    public Appointment findByIdentificationCode(String identificationCode){
+        return appointmentRepository.findByIdentificationCode(identificationCode)
+                .orElseThrow(() -> new NoSuchElementException());
+    }
+
+    public Appointment saveAppointment(Appointment appointment){
+        return appointmentRepository.save(appointment);
+    }
+}
