--- backend/WWMeet_backend/src/main/java/com/example/wwmeet_backend/dto/AppointmentRequestDto.java
@@ -1,20 +1,0 @@
-package com.example.wwmeet_backend.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
-
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class AppointmentRequestDto {
-    private Long id;
-    private String appointmentName;
-    private String appointmentPlace;
-    private String identificationCode;
-    private int peopleNum;
-    private LocalDateTime deadline;
-}
