--- backend/WWMeet_backend/src/main/java/com/example/wwmeet_backend/dto/ParticipantResponse.java
@@ -1,14 +1,0 @@
-package com.example.wwmeet_backend.dto;
-
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class ParticipantResponse {
-    private String appointmentIdentificationCode;
-    private String participantName;
-}
