--- backend/WWMeet_backend/src/main/java/com/example/wwmeet_backend/dto/PossibleScheduleResponseDto.java
@@ -1,30 +1,0 @@
-package com.example.wwmeet_backend.dto;
-
-import com.example.wwmeet_backend.domain.PossibleSchedule;
-import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
-import com.fasterxml.jackson.databind.annotation.JsonSerialize;
-import com.fasterxml.jackson.datatype.jsr310.deser.LocalDateTimeDeserializer;
-import com.fasterxml.jackson.datatype.jsr310.ser.LocalDateTimeSerializer;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class PossibleScheduleResponseDto {
-    @JsonSerialize(using = LocalDateTimeSerializer.class)
-    @JsonDeserialize(using = LocalDateTimeDeserializer.class)
-    private LocalDateTime startTime;
-
-    @JsonSerialize(using = LocalDateTimeSerializer.class)
-    @JsonDeserialize(using = LocalDateTimeDeserializer.class)
-    private LocalDateTime endTime;
-
-    public void setPossibleScheduleFrom(PossibleSchedule possibleSchedule) {
-        this.startTime = possibleSchedule.getStartTime();
-        this.endTime = possibleSchedule.getEndTime();
-    }
-}
