--- backend/WWMeet_backend/src/main/java/com/example/wwmeet_backend/dto/SaveAppointmentResponseDto.java
@@ -1,13 +1,0 @@
-package com.example.wwmeet_backend.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class SaveAppointmentResponseDto {
-    private AppointmentResponseDto appointmentRequestDto;
-    private String participantName;
-}
