--- backend/WWMeet_backend/src/main/java/com/example/wwmeet_backend/dto/VoteResponseDto.java
@@ -1,16 +1,0 @@
-package com.example.wwmeet_backend.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import java.util.List;
-
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class VoteResponseDto {
-    private Long appointmentId;
-    private Long participantId;
-    private List<PossibleScheduleResponseDto> possibleScheduleList;
-}
