--- backend/WWMeet_backend/src/main/java/com/example/wwmeet_backend/mapper/AppointmentMapper.java
@@ -1,15 +1,0 @@
-package com.example.wwmeet_backend.mapper;
-
-import com.example.wwmeet_backend.domain.Appointment;
-import com.example.wwmeet_backend.dto.AppointmentRequestDto;
-import com.example.wwmeet_backend.dto.AppointmentResponseDto;
-import org.mapstruct.Mapper;
-import org.springframework.context.annotation.Import;
-
-@Mapper(componentModel =  "spring")
-public interface AppointmentMapper {
-    AppointmentResponseDto toResponseDto(Appointment appointment);
-    Appointment responseDtoToEntity(AppointmentResponseDto appointmentFindDto);
-
-    Appointment requestDtoToEntity(AppointmentRequestDto appointmentRequestDto);
-}
