+++ backend/WWMeet_backend/src/main/java/com/example/wwmeet_backend/participant/dto/ParticipantResponse.java
@@ -1,0 +1,14 @@
+package com.example.wwmeet_backend.participant.dto;
+
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class ParticipantResponse {
+    private String appointmentIdentificationCode;
+    private String participantName;
+}
