+++ backend/WWMeet_backend/src/main/java/com/example/wwmeet_backend/possibleschedule/dto/PossibleScheduleResponseDto.java
@@ -1,0 +1,30 @@
+package com.example.wwmeet_backend.possibleschedule.dto;
+
+import com.example.wwmeet_backend.possibleschedule.domain.PossibleSchedule;
+import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
+import com.fasterxml.jackson.databind.annotation.JsonSerialize;
+import com.fasterxml.jackson.datatype.jsr310.deser.LocalDateTimeDeserializer;
+import com.fasterxml.jackson.datatype.jsr310.ser.LocalDateTimeSerializer;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.time.LocalDateTime;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class PossibleScheduleResponseDto {
+    @JsonSerialize(using = LocalDateTimeSerializer.class)
+    @JsonDeserialize(using = LocalDateTimeDeserializer.class)
+    private LocalDateTime startTime;
+
+    @JsonSerialize(using = LocalDateTimeSerializer.class)
+    @JsonDeserialize(using = LocalDateTimeDeserializer.class)
+    private LocalDateTime endTime;
+
+    public void setPossibleScheduleFrom(PossibleSchedule possibleSchedule) {
+        this.startTime = possibleSchedule.getStartTime();
+        this.endTime = possibleSchedule.getEndTime();
+    }
+}
