--- backend/WWMeet_backend/src/main/java/com/example/wwmeet_backend/repository/ParticipantRepository.java
@@ -1,8 +1,0 @@
-package com.example.wwmeet_backend.repository;
-
-import com.example.wwmeet_backend.domain.Participant;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ParticipantRepository extends JpaRepository<Participant, Long> {
-
-}
