--- backend/WWMeet_backend/src/main/java/com/example/wwmeet_backend/repository/VoteRepository.java
@@ -1,7 +1,0 @@
-package com.example.wwmeet_backend.repository;
-
-import com.example.wwmeet_backend.domain.Vote;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface VoteRepository extends JpaRepository<Vote, Long> {
-}
