+++ backend/WWMeet_backend/src/main/java/com/example/wwmeet_backend/vote/dto/VoteResponseDto.java
@@ -1,0 +1,18 @@
+package com.example.wwmeet_backend.vote.dto;
+
+import com.example.wwmeet_backend.possibleschedule.dto.PossibleScheduleResponseDto;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.util.List;
+
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class VoteResponseDto {
+    private Long appointmentId;
+    private Long participantId;
+    private List<PossibleScheduleResponseDto> possibleScheduleList;
+}
