+++ backend/WWMeet_backend/src/main/java/com/example/wwmeet_backend/vote/repository/VoteRepository.java
@@ -1,0 +1,7 @@
+package com.example.wwmeet_backend.vote.repository;
+
+import com.example.wwmeet_backend.vote.domain.Vote;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface VoteRepository extends JpaRepository<Vote, Long> {
+}
