+++ backend/WWMeet_backend/src/test/java/com/example/wwmeet_backend/service/AppointmentServiceTest.java
@@ -1,7 +1,8 @@
 package com.example.wwmeet_backend.service;
 
-import com.example.wwmeet_backend.domain.Appointment;
-import com.example.wwmeet_backend.repository.AppointmentRepository;
+import com.example.wwmeet_backend.appointment.domain.Appointment;
+import com.example.wwmeet_backend.appointment.repository.AppointmentRepository;
+import com.example.wwmeet_backend.appointment.service.AppointmentService;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
